fix: Repair request/query capture and overhaul Cache + Timeline UI
Request capture was silently broken: the request-correlation migration
had never run against production, so every entry insert failed and was
swallowed by a blanket catch in Monitor::flush(). Storage failures now
report through the app's exception handler instead of vanishing.

MarkControllerStart also never actually attached to routes — pushing
onto a middleware *group* array gets wiped whenever another package
re-syncs the Kernel's groups to the router afterwards. It now attaches
directly to the matched Route at RouteMatched time (Nightwatch's
approach), which can't be clobbered that way.

Other changes:
- Record duration to 2 decimal places end-to-end (recorders, Entry,
  Monitor, DatabaseStorage reads, migration widening the DB column).
- SlowQueries now counts every query executed inside a request (not
  just ones over threshold), so the Queries count and Request Detail
  timeline reflect reality; the dedicated Slow Queries digest still
  filters to just the slow ones via a subtype flag.
- Redesign the Request Detail timeline to feel like Nightwatch: a
  divided, indented event tree, a zoom range slider (replacing +/-
  buttons) that preserves scroll position, click-and-drag panning, and
  hover tooltips showing untruncated SQL/cache keys.
- Add a proper Cache page: Events/Failures charts (hits, misses,
  writes, deletes, write/forget failures) and a sortable per-key
  table, mirroring Nightwatch's Cache view. Wires up KeyWriteFailed/
  KeyForgetFailed, which the recorder never listened to before.
- Standardize chart card heights across the dashboard pages.

+++ src/Support/Timeline.php
@@ -35,7 +35,7 @@ class Timeline
      */
     public static function build(object $root, Collection $children): array
     {
-        $duration = (int) ($root->duration ?? 0);
+        $duration = (float) ($root->duration ?? 0);
 
         $requestEntry = new TimelineEntry(id: 'request', type: 'request', label: $root->key ?? 'Request', start: 0, duration: $duration);
 
@@ -86,7 +86,7 @@ class Timeline
         $map = self::EVENT_TYPES[$row->type] ?? ['type' => $row->type, 'label' => ucfirst($row->type)];
 
         $start = max(0, (int) ($row->start_offset ?? 0));
-        $duration = max(0, (int) ($row->duration ?? 0));
+        $duration = max(0.0, (float) ($row->duration ?? 0));
 
         return new TimelineEntry(
             id: (string) $row->id,
